+++ src/c/appendix/math.h
@@ -1,3 +1,5 @@
 #pragma once
 
-void min_max(int *array, int n, int *min, int *max);
+#include <pebble.h>
+
+void min_max(int16_t *array, int n, int *min, int *max);
